requirements.txt
flask~=1.1.2
flask_restful
flask_script
python-dotenv~=0.13.0
flask-sqlalchemy
pymysql
python-dateutil
flask_marshmallow
SQLAlchemy~=1.3.17